feat: implementa estrutura base do CRUD de lembretes
- Adiciona GlobalExceptionHandler e ResourceNotFoundException

- Atualiza dependências no pom.xml

- Ajusta regras do .gitignore

- Refatora Controller, Service e Repository

+++ src/main/java/org/example/reminder/service/ReminderService.java
@@ -1,7 +1,10 @@
 package org.example.reminder.service;
+
 import org.example.reminder.dto.ReminderRequestDTO;
 import org.example.reminder.entity.Reminder;
+import org.example.reminder.exception.ResourceNotFoundException;
 import org.example.reminder.repository.ReminderRepository;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,19 +29,27 @@ public class ReminderService {
         return reminderRepository.save(entity);
     }
 
-    public List<Reminder> getAll() {
-        return reminderRepository.findAll();
+    public List<Reminder> findAll() {
+        return reminderRepository.findAll(Sort.by("dueDate").ascending());
     }
 
     public void delete(Long id) {
+        if (!reminderRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Lembrete não encontrado com id: " + id);
+        }
         reminderRepository.deleteById(id);
     }
 
     public Reminder update(Long id, ReminderRequestDTO dto) {
-        Reminder entity = reminderRepository.findById(id).get();
-        entity.setTitle( dto.title());
+        Reminder entity = reminderRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Lembrete não encontrado com id: " + id));
+        entity.setTitle(dto.title());
         entity.setDescription(dto.description());
         entity.setDueDate(dto.dueDate());
+
+        if (dto.completed() != null) {
+            entity.setCompleted(dto.completed());
+        }
 
         return reminderRepository.save(entity);
     }
